Feature/architecture adjustement (#12)
FEATURE removed ArchViewModel, as it's useless. We want a "dumb"
ViewModel only responsible to map state from Interactor to object being
consume by the UI

ArchViewHolder now need a mainCoroutineDispatcher to be created,
bindOutputs() is now a suspend function and his inside a Coroutine with
the mainCoroutineDispatcher

fixed a potential leak with Coroutine, added a job to create a
uiScope. He will be cancel when onViewDetachedFromWindow is called

removed Dagger from arch library